+++ Drivers/template/SQLGetTypeInfo.c
@@ -10,6 +10,7 @@
  *
  **********************************************************************/
 
+#include <config.h>
 #include "driver.h"
 
 SQLRETURN SQLGetTypeInfo( SQLHSTMT        hDrvStmt,
@@ -21,7 +22,7 @@ SQLRETURN SQLGetTypeInfo( SQLHSTMT        hDrvStmt,
 	if ( NULL == hStmt )
 		return SQL_INVALID_HANDLE;
 
-	sprintf((char*) hStmt->szSqlMsg, "hStmt = $%08lX", hStmt );
+	sprintf((char*) hStmt->szSqlMsg, "hStmt = $%08lX", (long)hStmt );
 	logPushMsg( hStmt->hLog, __FILE__, __FILE__, __LINE__, LOG_WARNING, LOG_WARNING,(char*) hStmt->szSqlMsg );
 
     /************************
